Test: Created separate header files for each different test
-Modular approach keeps main entry point clean
